move output functions to report
