+++ src/sqlfmt/report.py
@@ -3,9 +3,22 @@ from dataclasses import dataclass
 from pathlib import Path
 from typing import List, Optional
 
+import click
+
 from sqlfmt.exception import SqlfmtError
 from sqlfmt.mode import Mode
-from sqlfmt.utils import style_output
+
+
+def style_output(
+    msg: str, fg: Optional[str] = None, bg: Optional[str] = None, bold: bool = False
+) -> str:
+    # see https://click.palletsprojects.com/en/8.0.x/api/?highlight=style#click.style
+    s: str = click.style(msg, fg=fg, bg=bg, bold=bold)
+    return s
+
+
+def display_output(msg: str, err: bool = True) -> None:
+    click.echo(msg, err=err)
 
 
 @dataclass
